moved interceptor settings provider to class
